+++ Domain/Entities/MovimientoFinanciero.cs
@@ -3,11 +3,11 @@ using System;
 
 namespace Domain.Entities
 {
-    public class MovimientoFinanciero : Entity<int>
-    {
-        public ServicioFinanciero CuentaBancaria { get; set; }
-        public decimal ValorRetiro { get; set; }
-        public decimal ValorConsignacion { get; set; }
-        public DateTime FechaMovimiento { get; set; }
-    }
+	public class MovimientoFinanciero : Entity<int>
+	{
+		public ServicioFinanciero CuentaBancaria { get; set; }
+		public double ValorRetiro { get; set; }
+		public double ValorConsignacion { get; set; }
+		public DateTime FechaMovimiento { get; set; }
+	}
 }
